contacts and user endpoints update

+++ routes/api/contacts.js
@@ -3,6 +3,8 @@ const express = require("express");
 const router = express.Router();
 
 const Contact = require("../../models/contact");
+const User = require("../../models/user");
+
 const { boolean } = require("joi");
 
 const auth = require("../../config/authorization")
@@ -12,8 +14,35 @@ const auth = require("../../config/authorization")
 
 router.get("/", auth, async (req, res, next) => {
   try {
-    const data = await Contact.find();
-    res.json(data);
+    const page = parseInt(req.query.page) || 1;
+    const limit = parseInt(req.query.limit) || 20;
+    const token = req.headers.authorization.slice(7)
+
+    const skip = (page - 1) * limit;
+
+    const user = await User.findOne({ token });
+
+    if (!user) {
+      return res.status(401).json({
+        status: "Unauthorized",
+        code: 401,
+        message: "Invalid token",
+      });
+    }
+
+    const filter = {};
+
+    if (req.query.favorite) {
+      filter.favorite = req.query.favorite === "true";
+    }
+
+    const contacts = await Contact.find({ owner: user._id, ...filter }).skip(skip).limit(limit)
+
+    res.json({
+      page,
+      limit,
+      contacts,
+    });
   } catch (error) {
     res.status(404).json({
       status: "Not found",
@@ -25,7 +54,19 @@ router.get("/", auth, async (req, res, next) => {
 
 router.get("/:contactId", auth, async (req, res, next) => {
   try {
-    data = await Contact.findOne({ _id: req.params.contactId });
+    const token = req.headers.authorization.slice(7)
+    const user = await User.findOne({ token });
+    console.log(user)
+
+    if (!user) {
+      return res.status(401).json({
+        status: "Unauthorized",
+        code: 401,
+        message: "Invalid token",
+      });
+    }
+
+    data = await Contact.findOne({ _id: req.params.contactId,  owner: user._id });
 
     return res.json(data);
   } catch (error) {
@@ -38,18 +79,41 @@ router.get("/:contactId", auth, async (req, res, next) => {
 });
 
 router.post("/", auth, async (req, res, next) => {
-  try {
+
+  const token = req.headers.authorization.slice(7)
+  const user = await User.findOne({ token });
+
+  if (!user) {
+    return res.status(401).json({
+      status: "Unauthorized",
+      code: 401,
+      message: "Invalid token",
+    });
+  }
+  
+   try {
     const data = await Contact.create(req.body);
     console.log(data)
     res.json(data);
   } catch (error) {
     next(error);
-  }
+  } 
 });
 
 router.delete("/:contactId", auth, async (req, res, next) => {
   try {
-    data = await Contact.findByIdAndRemove({ _id: req.params.contactId });
+    const token = req.headers.authorization.slice(7)
+    const user = await User.findOne({ token });
+  
+    if (!user) {
+      return res.status(401).json({
+        status: "Unauthorized",
+        code: 401,
+        message: "Invalid token",
+      });
+    }
+
+    data = await Contact.findByIdAndRemove({ _id: req.params.contactId, owner: user._id });
     if (data) {
       return res.json({
         message: `Contact with ${req.params.contactId} deleted`,
@@ -63,13 +127,24 @@ router.delete("/:contactId", auth, async (req, res, next) => {
 
 router.put("/:contactId", auth, async (req, res, next) => {
   try {
+    const token = req.headers.authorization.slice(7)
+    const user = await User.findOne({ token });
+  
+    if (!user) {
+      return res.status(401).json({
+        status: "Unauthorized",
+        code: 401,
+        message: "Invalid token",
+      });
+    }
+
     data = await Contact.findByIdAndUpdate(
       { _id: req.params.contactId },
       req.body,
-      { new: true }
+      { new: true },
+      { owner: user._id },
     );
 
-    console.log(data);
     if (data) {
       return res.json(data);
     } else {
@@ -89,9 +164,21 @@ router.put("/:contactId", auth, async (req, res, next) => {
 
 router.patch("/:contactId/favorite", auth, async (req, res, next) => {
   try {
+    const token = req.headers.authorization.slice(7)
+    const user = await User.findOne({ token });
+  
+    if (!user) {
+      return res.status(401).json({
+        status: "Unauthorized",
+        code: 401,
+        message: "Invalid token",
+      });
+    }
+
     data = await Contact.findByIdAndUpdate(
       { _id: req.params.contactId },
-      req.body
+      req.body,
+      { owner: user._id },
     );
     if (!req.body.favorite === boolean) {
       return res.status(400).send({ "message": "missing field favorite" });
